requirements.txt
torch
numpy
matplotlib
tqdm
pyyaml
tensorboardx